+++ ClientApp/src/app/app-routing.module.ts
@@ -1,0 +1,23 @@
+﻿import { NgModule } from '@angular/core';
+import { Routes, RouterModule, PreloadAllModules } from '@angular/router';
+
+import { HomeComponent } from './home/home.component';
+
+const routes: Routes = [
+    { path: '', redirectTo: 'home', pathMatch: 'full' },
+    { path: 'home', component: HomeComponent },
+    { path: 'account', loadChildren: './account/account.module#AccountModule' },
+    { path: 'resources', loadChildren: './resources/resources.module#ResourcesModule' },
+    { path: 'dashboard', loadChildren: './dashboard/dashboard.module#DashboardModule' },
+    { path: '**', redirectTo: 'home' }
+];
+
+@NgModule({
+    imports: [
+        RouterModule.forRoot(routes, {
+            preloadingStrategy: PreloadAllModules
+        })
+    ],
+    exports: [RouterModule]
+})
+export class AppRoutingModule { }