+++ ClientApp/src/environments/environment.staging.ts
@@ -1,0 +1,4 @@
+export const environment: any = {
+    production: false,
+    staging: true
+};