+++ Models/AccountViewModels/CreateViewModel.cs
@@ -1,8 +1,5 @@
 ﻿namespace AngularSPAWebAPI.Models.AccountViewModels
 {
-    /// <summary>
-    /// Class required to create a new user.
-    /// </summary>
     public class CreateViewModel
     {
         public string username { get; set; }