+++ src/environments/environment.hmr.ts
@@ -1,5 +1,0 @@
-export const environment: any = {
-    production: false,
-    staging: false,
-    hmr: true
-};