+++ src/environments/environment.prod.ts
@@ -1,5 +1,0 @@
-export const environment: any = {
-    production: true,
-    staging: false,
-    hmr: false
-};