+++ src/environments/environment.staging.ts
@@ -1,5 +1,0 @@
-export const environment: any = {
-    production: false,
-    staging: true,
-    hmr: false
-};